# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/pullRequest/create": {
            "post": {
                "description": "create new pr and assign reviewers automatically",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "pull requests"
                ],
                "summary": "Create new Pull Request",
                "parameters": [
                    {
                        "description": "PR DATA",
                        "name": "query",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.PullRequestQuery"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/dto.PrResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/pullRequest/merge": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "pull requests"
                ],
                "summary": "Merge Pull Request",
                "parameters": [
                    {
                        "description": "PR ID",
                        "name": "query",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.PullRequestIdQuery"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dto.PrMergedResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/pullRequest/reassign": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "pull requests"
                ],
                "summary": "Reassign reviewer Pull Request",
                "parameters": [
                    {
                        "description": "Pr id, old reviewer id",
                        "name": "query",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.PrReassignQuery"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dto.PrReassignResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/stat/pull_request/reviewers": {
            "get": {
                "description": "get pull requests with reviewers and their number",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "statistics"
                ],
                "summary": "get prs with assigned reviewers",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.PrReviewersCount"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/stat/users/reviews": {
            "get": {
                "description": "get users and in how many pull requests they are reviewers",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "statistics"
                ],
                "summary": "get users and number of reviews",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.UserReviewsCount"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/team/add": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "teams"
                ],
                "summary": "add new team",
                "parameters": [
                    {
                        "description": "team object: team_name {members}",
                        "name": "team",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Team"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/model.Team"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/team/get": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "teams"
                ],
                "summary": "get existing team",
                "parameters": [
                    {
                        "type": "string",
                        "description": "team_name",
                        "name": "team_name",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Team"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/team/kill": {
            "post": {
                "description": "set users status to not active by a given team name",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "teams"
                ],
                "summary": "deactivate all users in team",
                "parameters": [
                    {
                        "description": "team_name",
                        "name": "team_name",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.TeamName"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Team"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/users/getReview": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "get prs where user is reviewer",
                "parameters": [
                    {
                        "type": "string",
                        "description": "user id",
                        "name": "user_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.UserPrsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/users/setIsActive": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "users"
                ],
                "summary": "set user is active status",
                "parameters": [
                    {
                        "description": "user id, status",
                        "name": "query",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.StatusQuery"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.UserResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "dto.PrMerged": {
            "type": "object",
            "properties": {
                "assigned_reviewers": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "author_id": {
                    "type": "string"
                },
                "merged_at": {
                    "type": "string"
                },
                "pull_request_id": {
                    "type": "string"
                },
                "pull_request_name": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/model.PRstatus"
                }
            }
        },
        "dto.PrMergedResponse": {
            "type": "object",
            "properties": {
                "pr": {
                    "$ref": "#/definitions/dto.PrMerged"
                }
            }
        },
        "dto.PrReassignQuery": {
            "type": "object",
            "properties": {
                "old_reviewer_id": {
                    "type": "string"
                },
                "pull_request_id": {
                    "type": "string"
                }
            }
        },
        "dto.PrReassignResponse": {
            "type": "object",
            "properties": {
                "pr": {
                    "$ref": "#/definitions/dto.PrResponse"
                },
                "replaced_by": {
                    "type": "string"
                }
            }
        },
        "dto.PrResponse": {
            "type": "object",
            "properties": {
                "assigned_reviewers": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "author_id": {
                    "type": "string"
                },
                "pull_request_id": {
                    "type": "string"
                },
                "pull_request_name": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/model.PRstatus"
                }
            }
        },
        "dto.PullRequestIdQuery": {
            "type": "object",
            "properties": {
                "pull_request_id": {
                    "type": "string"
                }
            }
        },
        "dto.PullRequestQuery": {
            "type": "object",
            "properties": {
                "author_id": {
                    "type": "string"
                },
                "pull_request_id": {
                    "type": "string"
                },
                "pull_request_name": {
                    "type": "string"
                }
            }
        },
        "dto.StatusQuery": {
            "type": "object",
            "properties": {
                "is_active": {
                    "type": "boolean"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "dto.TeamName": {
            "type": "object",
            "properties": {
                "team_name": {
                    "type": "string"
                }
            }
        },
        "dto.UserPrsResponse": {
            "type": "object",
            "properties": {
                "pull_requests": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.PullRequestShort"
                    }
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "dto.UserResponse": {
            "type": "object",
            "properties": {
                "user": {
                    "$ref": "#/definitions/model.User"
                }
            }
        },
        "model.CustomError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "message": {
                    "$ref": "#/definitions/model.ErrCode"
                }
            }
        },
        "model.ErrCode": {
            "type": "string",
            "enum": [
                "TEAM_EXISTS",
                "PR_EXISTS",
                "PR_MERGED",
                "NOT_ASSIGNED",
                "NO_CANDIDATE",
                "NOT_FOUND",
                "INTERNAL_ERROR"
            ],
            "x-enum-varnames": [
                "TEAM_EXISTS",
                "PR_EXISTS",
                "PR_MERGED",
                "NOT_ASSIGNED",
                "NO_CANDIDATE",
                "NOT_FOUND",
                "INTERNAL_ERROR"
            ]
        },
        "model.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/model.CustomError"
                }
            }
        },
        "model.PRstatus": {
            "type": "string",
            "enum": [
                "created",
                "merged"
            ],
            "x-enum-varnames": [
                "CREATED",
                "MERGED"
            ]
        },
        "model.PrReviewersCount": {
            "type": "object",
            "properties": {
                "pull_request": {
                    "$ref": "#/definitions/model.PullRequestShort"
                },
                "reviewers_count": {
                    "type": "integer"
                }
            }
        },
        "model.PullRequestShort": {
            "type": "object",
            "properties": {
                "author_id": {
                    "type": "string"
                },
                "pull_request_id": {
                    "type": "string"
                },
                "pull_request_name": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/model.PRstatus"
                }
            }
        },
        "model.Team": {
            "type": "object",
            "properties": {
                "members": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.TeamMember"
                    }
                },
                "team_name": {
                    "type": "string"
                }
            }
        },
        "model.TeamMember": {
            "type": "object",
            "properties": {
                "is_active": {
                    "type": "boolean"
                },
                "user_id": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "model.User": {
            "type": "object",
            "properties": {
                "is_active": {
                    "type": "boolean"
                },
                "team_name": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "model.UserReviewsCount": {
            "type": "object",
            "properties": {
                "reviews_count": {
                    "type": "integer"
                },
                "user_id": {
                    "$ref": "#/definitions/model.User"
                }
            }
        }
    }
}